feat(indexer): extract SorobanRpcClient and LedgerFollower into own modules
Resolves #188 and #189 by splitting the monolithic indexer.rs into three
focused modules so each component is independently testable and named:

- soroban_rpc: SorobanRpcClient with configurable retry/backoff (issue #188)
- ledger_follower: LedgerFollower with Postgres checkpointing (issue #189)
- indexer_metrics: shared atomic counters, decoupled from both modules
- indexer: thin orchestration layer; re-exports metrics() for health routes

Also adds the missing indexed_disputes migration that the DisputeOpened
event handler (merged in #329) writes to.
