+++ backend/src/main.rs
@@ -8,10 +8,13 @@ mod db;
 mod env_config;
 mod error;
 mod indexer;
+mod indexer_metrics;
+mod ledger_follower;
 mod middleware;
 mod models;
 mod routes;
 mod services;
+mod soroban_rpc;
 mod tx_metadata_cache;
 mod tx_queue;
 mod worker;
